String splitter
Basically, what we gonna do is, we convert float into string, and send
the string to the slave. Slave gats the charactor then convert back to
String, then we can use , String indexOf(), get the location of the mark
I made. We use those index to get the data we want using, subString(),
function.